fix(test): sizeof_member_via_null_cast widens `long` to `long long`
`long` is 4 bytes on Windows LLP64, so the nested-member width
check fired on the windows-x64 and windows-arm64 lanes. Switch
the deepest field to `long long` (8 bytes on every supported
target) and note the LP64-vs-LLP64 split next to the field
declaration so a later sweep doesn't reintroduce the mismatch.

+++ tests/fixtures/c/sizeof_member_via_null_cast.c
@@ -18,7 +18,9 @@ struct Stab_Sym {
 
 struct Nested {
     struct {
-        long deep;
+        // long long is 8 bytes on every supported target;
+        // `long` would be 4 bytes on Windows LLP64.
+        long long deep;
     } inner;
 };
